\name{inferDistribution-package}
\alias{inferDistribution-package}
\alias{inferDistribution}
\docType{package}
\title{
\packageTitle{inferDistribution}
}
\description{
\packageDescription{inferDistribution}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{inferDistribution}
\packageIndices{inferDistribution}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{inferDistribution}

Maintainer: \packageMaintainer{inferDistribution}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
